+++ src/Manager/Environment/segger.test.js
@@ -34,6 +34,15 @@
  * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+/* eslint-disable import/first */
+jest.mock('electron', () => ({
+    remote: {
+        require: () => ({
+            exec: jest.fn(),
+        }),
+    },
+}));
+
 import path from 'path';
 import { updateSettingsXml, userSettings } from './segger';