+++ src/persistentStore.js
@@ -52,15 +52,19 @@ const defaultInstallDir = {
 export const persistedInstallDir = () => store.get('installDir', defaultInstallDir);
 export const setPersistedInstallDir = dir => store.set('installDir', dir);
 
-const defaultIndexJson = {
+const indexJson = {
     win32: 'index.json',
     darwin: 'index-mac.json',
     linux: 'index-linux.json',
 }[process.platform];
 
-export const toolchainIndexUrl = () => store.get('toolchainIndexUrl',
-    `https://developer.nordicsemi.com/.pc-tools/toolchain/${defaultIndexJson}`);
+export const toolchainIndexUrl = () => {
+    const value = store.get('toolchainIndexUrl',
+        'https://developer.nordicsemi.com/.pc-tools/toolchain');
+    return `${value.replace(/\/index.*.json$/, '')}/${indexJson}`;
+};
 export const toolchainUrl = name => `${path.dirname(toolchainIndexUrl())}/${name}`;
+export const setToolchainIndexUrl = value => store.set('toolchainIndexUrl', value);
 
 export const persistedShowMaster = () => store.get('showMaster', false);
 export const setPersistedShowMaster = visible => store.set('showMaster', visible);