+++ Changelog.md
@@ -1,3 +1,7 @@
+## Unreleased
+- Feature: settings to change toolchain source
+- Linux toolchain support
+
 ## Version 0.7.1
 - Fixed update SDK/toolchain actions